+++ examples/certilizer-cli.sh
@@ -21,3 +21,7 @@ certilizer --conf-file certilizer.yaml
 printf "\n\n========================================\n"
 printf "Run command with specified config file, output format, and output file:\n"
 certilizer --conf-file certilizer.yaml --out-format html --out-file ../stage/example-output.html
+
+printf "\n\n========================================\n"
+printf "Run command with specified maximum column size and expiry threshold in 180 days:\n"
+certilizer --conf-file certilizer.yaml --out-format html --max-col-size 20 --expiry-threshold-in-days 180 --out-file ../stage/example-output-180days.html